requirements.txt
Flask==2.0.2
requests==2.22.0
easyocr==1.4.1
Flask_RESTful==0.3.9
numpy==1.19.4
Pillow==8.3.2
